// src/gamification/enums/badge-criteria-type.enum.ts
export enum BadgeCriteriaType {
  COURSES_COMPLETED = 'COURSES_COMPLETED',
  LESSONS_WATCHED = 'LESSONS_WATCHED',
  LEARNING_STREAK_DAYS = 'LEARNING_STREAK_DAYS',
  ASSESSMENT_PERFECT_SCORE = 'ASSESSMENT_PERFECT_SCORE',
  ASSESSMENTS_PASSED = 'ASSESSMENTS_PASSED',
  POINTS_REACHED = 'POINTS_REACHED',
  LEVEL_REACHED = 'LEVEL_REACHED',
  REVIEWS_WRITTEN = 'REVIEWS_WRITTEN',
  COURSES_CREATED = 'COURSES_CREATED',
}